raspi-controller/requirements.txt
configparser
asyncio
websockets
pyserial